refactor: movendo ChatProxyClient para pacote client
